feat(perf): Add tooltip to closeContainer for widgets
This cleans up some styling code around the selectable list items, and separates close container into a re-usable container for all the mini list widgets to use. The new component also adds the tooltip explaining what clicking the 'x' does so people don't think it's a destructive operation.
