+++ static/app/views/performance/landing/widgets/components/selectableList.tsx
@@ -1,7 +1,12 @@
 import React, {ReactNode} from 'react';
 import styled from '@emotion/styled';
 
+import EmptyStateWarning from 'app/components/emptyStateWarning';
+import Link from 'app/components/links/link';
 import Radio from 'app/components/radio';
+import Tooltip from 'app/components/tooltip';
+import {IconClose} from 'app/icons';
+import {t} from 'app/locale';
 import space from 'app/styles/space';
 import {RadioLineItem} from 'app/views/settings/components/forms/controls/radioGroup';
 
@@ -50,6 +55,57 @@ function SelectableItem({
 
 export const RightAlignedCell = styled('div')`
   text-align: right;
+`;
+
+export const Subtitle = styled('span')`
+  color: ${p => p.theme.gray300};
+  font-size: ${p => p.theme.fontSizeMedium};
+`;
+export const GrowLink = styled(Link)`
+  flex-grow: 1;
+`;
+
+export const WidgetEmptyStateWarning = () => {
+  return <StyledEmptyStateWarning small>{t('No results')}</StyledEmptyStateWarning>;
+};
+
+export function ListClose(props: {
+  onClick: () => void;
+  setSelectListIndex: (n: number) => void;
+}) {
+  return (
+    <CloseContainer>
+      <Tooltip title={t('Exclude this transaction from the search filter.')}>
+        <StyledIconClose
+          onClick={() => {
+            props.onClick();
+            props.setSelectListIndex(0);
+          }}
+        />
+      </Tooltip>
+    </CloseContainer>
+  );
+}
+
+const CloseContainer = styled('div')`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-left: ${space(1)};
+`;
+
+const StyledIconClose = styled(IconClose)`
+  cursor: pointer;
+  color: ${p => p.theme.gray200};
+
+  &:hover {
+    color: ${p => p.theme.gray300};
+  }
+`;
+
+const StyledEmptyStateWarning = styled(EmptyStateWarning)`
+  min-height: 300px;
+  justify-content: center;
 `;
 
 const ListItemContainer = styled('div')`
